requirements.txt
fastapi
uvicorn[standard]
uform
usearch
torch
numpy
Pillow
python-multipart
